requirements.txt
# --------- python --------- #
numpy<2
matplotlib
pandas
scipy
ipykernel
Pillow

# --------- pytorch --------- #
# torch
# torchvision
pytorch-lightning==2.2.3
jsonargparse[signatures]
torchmetrics

# --------- loggers --------- #
wandb

# --------- custom --------- #
opencv-python
transformers
git+https://github.com/openai/CLIP.git
git+https://github.com/facebookresearch/pytorchvideo.git

# --------- misc --------- #
huggingface_hub[cli]
tqdm
pyrootutils
python-dotenv
rich
